Add basic player, mobile controller, and input
The SixDoFController is used by the player to both move and rotate. It
works with mouse & keyboard, controller, and touch controls through the
MobileControllerScript.

The MobileControllerScript is a generalised script used to create and
manage touch joysticks and buttons for mobile. It is given to an empty
object called "Mobile Controller" where other scripts expect to find it.

All of the currently needed input axes have been set up, allowing both
mouse & keyboard and controller input to be handled correctly.