Add basic projectile system
The basic system consists of the projectile prefab and a ShootScript.

The projectile prefab has a ProjectileScript that will attempt to deal
damage upon colliding with something, and is virtual, allowing other
projectiles to be created for it. Upon hitting something, it also
instantiates another prefab as an explosion, which should then destroy
itself.

The ShootScript is effectively the gun that creates the projectile. It
is used by another script, which calls its Shoot method to attempt to
fire a projectile. It has a bool for semi auto mode, but it should be
implemented by the script calling Shoot, and is not used by ShootScript.