Add basic interaction system
The basic system has the PlayerInteract script on the player, with
interactable objects having a script inheriting from Interactable,
overriding Interaction. PlayerInteract uses an event that Interactable
adds a listener to while the player is within its trigger.

The mobile controls are currently broken, as the interaction button is
the second button, with shooting having the first, which isnt
implemented yet.