Add save game system
The SaveGameScript is used to save game data held in its public struct
SaveData. This data is saved and loaded from a file located at a
persistant directory. It has an event that is triggered when its data
needs to be updated before saving, avalable in the inspector.

Save points are objects that act as checkpoints in a level. They only
have a SavePointScript, deriving from Interactable, that will update the
save's savePoint and save the current state to file.

PlayerManager, which has not yet been added, will set the player's
position to the save point with the same index as the save file.