Add shield and win zone
The shield is just anobject with a trigger and "Shield" tag, but is also
given a custom shader with backface culling off so the inside is
rendered. ProjectileScript checks for triggers with this tag to collide
with, so it doesnt need to do anything.

The win zone is just a trigger with a WinZoneScript, without even a mesh
or mesh renderer, because it doesnt need to be visible. When the player
enters the trigger, it will cause a game over with the win state.