Add turret
The turret has a TurretScript that controls it, alog with a requiered
ShootScript. It has multiple children, but they can mostly be ignored.
It will cast a ray toward its target (set to the player), and if it
hits, aim the barrel toward them and start shooting, until it looses
sight of them. In order to not have the ray hitting everything, a layer
mask is used, with a new layer being added to objects it should be able
to see through, such as the shield.