Add options menu
The options menu is accesable from the main menu, and provides two
things: settings and controls. Settings can be set using sliders and
toggles to set PlayerPrefs used in the game. Controls can be viewed by
pressing a button, displaying the appropriately labeled control scheme.

The settings are all operational without issue. The controls sellection
works, however the controls displayed are only placeholder and need to
be changed.