Fix button text and notification
The text in buttons was set to best fit while thier size filled the
button, resulting in the text commpleatly taking up the space. Now they
are slightly smaller, making them look much nicer.

The notification text on the HUD wouldnt show up, as the GameObject was
inactive while the script only enables the Text component. Now only the
component is set as inactive, fixing the issue.