Move player, canvas, etc to BaseGameScene
With a single level, the player and UI just exists, but when more scenes
begin to be created, it becomes an issue with having to set them up
again each time. To resolve this, the player, canvas, EventSystem, and
Mobile Controller has been moved to a seperate scene, BaseGameScene.

This is only managed by SceneTransitionScript, where it will load the
correct level in single mode, then the base scene in additive.

When editing levels, the base scene should be opened alongside it for
testing.