Fix respawn and restart buttons
When implementing BaseGameScene, only SceneTransitionScript would load
it, which resulted in starting the level working fine. When the level is
reloaded, however, it is actualy just reopened in single, closing all
other scenes, so the player would not exist anymore.

Now these buttons also open the base scene, fixing the issue.