Change MobileControllerScript to use dictionary
The public lists used to store the joystick and button input was kept in
order of their creation. This posed a problem in cases where multiple
scripts have to create a button, as the index becomes dependant on their
order of execution.

Now that these lists have been changed to dictionaries, a string is used
to access input. The key is passed to the function used to create the
joystick/button.