Add enemy base class, KillEnemiesConditionEvent
TurretScript now inherits from EnemyScript, which only implements
DealDamage. The reason for this addition is to include an event that is
invoked when the enemy is killed. Instead of a unity event, EnemyScript
uses a C# event.

KillEnemiesConditionEvent is a script that takes two things; a list of
EnemyScripts which is attaches listeners to the die evens of, and a
unity event that is invoked when all the enemies are killed.

Using KillEnemiesConditionEvent with the new ObjectMoveScript, it is
possible to mkae 'doors' that open after killing all enemies in a room,
which is exactly what has been added to the three levels, along with
some minor geometry fixes that were missed.