Fix door in level 3, move arrow
One of the doors in level 3 wouldmove up twice in responce to two
diferent kill conditions, as intended. However, the second movement
moved it into another room, which was not realised. This has now been
changed, and it moves sideways on the second event.

An arrow was in a wall, and couldnt be seen after all the levels were
snapped to grid. It has now been put back where it was.